requirements.txt
pip>=21.3.1
wheel~=0.37.1
rsa~=4.8
pyasn1~=0.4.8
pytz~=2022.1
PyYAML~=6.0
attrs~=22.1.0
MarkupSafe~=2.1.1
Werkzeug~=2.2.1
Jinja2~=3.1.2
click~=8.1.3
itsdangerous~=2.1.2
numpy~=1.23.1
setuptools~=60.2.0
six~=1.16.0
cachetools~=5.2.0
certifi~=2022.6.15
urllib3~=1.26.10
requests~=2.28.1
pandas~=1.4.3
python-dateutil~=2.8.2
SQLAlchemy~=1.4.39
greenlet~=1.1.2
oauthlib~=3.2.0
idna~=3.3
Flask~=2.1.3
connexion~=2.14.0
clickclick~=20.10.2
jsonschema~=4.9.0
inflection~=0.5.1
packaging~=21.3
pyparsing~=3.0.9
pyrsistent~=0.18.1
marshmallow~=3.17.0
gspread~=5.4.0
python-dotenv~=0.20.0
pymongo>=4.2.0
python-telegram-bot>= 13.13